Check :pedantic in project to warn instead of fail
Hook get-dependencies instead of in pomegranate.  Then we can
get the project map and check the :pedantic key.  If it is :warn
then print a warning, otherwise fail.

Inject a middleware that looks at the profiles for :dependencies
being added by the repl task.  If found then default to warning.